Add Authorize

+++ src/AuctionService/Controllers/AuctionsController.cs
@@ -5,6 +5,7 @@ using AutoMapper;
 using AutoMapper.QueryableExtensions;
 using Contracts;
 using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -43,13 +44,14 @@ public class AuctionsController(AuctionDbContext context, IMapper mapper, IPubli
         return mapper.Map<AuctionDto>(auction);
     }
 
+    [Authorize]
     [HttpPost]
     public async Task<ActionResult<AuctionDto>> CreateAuction([FromBody] CreateAuctionDto auctionDto)
     {
         var auction = mapper.Map<Auction>(auctionDto);
 
-        // TODO: add current user as seller
         auction.Seller = "test";
+        auction.Seller = User.Identity?.Name ?? throw new InvalidOperationException("User not found");
 
         context.Auctions.Add(auction);
 
@@ -64,6 +66,7 @@ public class AuctionsController(AuctionDbContext context, IMapper mapper, IPubli
         return CreatedAtAction(nameof(GetAuction), new { id = auction.Id }, newAuction);
     }
 
+    [Authorize]
     [HttpPut("{id}")]
     public async Task<ActionResult<AuctionDto>> UpdateAuction(Guid id, [FromBody] UpdateAuctionDto auctionDto)
     {
@@ -76,7 +79,7 @@ public class AuctionsController(AuctionDbContext context, IMapper mapper, IPubli
             return NotFound();
         }
 
-        // TODO: check seller is the same as current user
+        if (auction.Seller != User.Identity?.Name) return Forbid();
 
         auction.Item.Name = auctionDto.Name ?? auction.Item.Name;
         auction.Item.Genre = auctionDto.Genre ?? auction.Item.Genre;
@@ -94,6 +97,7 @@ public class AuctionsController(AuctionDbContext context, IMapper mapper, IPubli
         return Ok(mapper.Map<AuctionDto>(auction));
     }
 
+    [Authorize]
     [HttpDelete("{id}")]
     public async Task<ActionResult> DeleteAuction(Guid id) 
     {
@@ -103,7 +107,7 @@ public class AuctionsController(AuctionDbContext context, IMapper mapper, IPubli
             return NotFound();
         }
 
-         // TODO: check seller is the same as current user
+        if (auction.Seller != User.Identity?.Name) return Forbid();
         context.Auctions.Remove(auction);
 
         await publishEndpoint.Publish<AuctionDeleted>(new {Id = auction.Id.ToString()});
